+++ profilemanager.h
@@ -16,7 +16,7 @@ class ProfileInfo : public QObject {
     Q_PROPERTY(bool windowCustomSize MEMBER windowCustomSize NOTIFY changed)
     Q_PROPERTY(int windowWidth MEMBER windowWidth NOTIFY changed)
     Q_PROPERTY(int windowHeight MEMBER windowHeight NOTIFY changed)
-    Q_PROPERTY(int pixelScale MEMBER pixelScale NOTIFY changed)
+    Q_PROPERTY(float pixelScale MEMBER pixelScale NOTIFY changed)
 
     ProfileManager* manager;
 
@@ -36,7 +36,7 @@ public:
     bool windowCustomSize = false;
     int windowWidth = 720;
     int windowHeight = 480;
-    int pixelScale = 2;
+    float pixelScale = 2;
 
 public slots:
     void setName(QString const& newName);
